requirements.txt
ftfy
html2text
pyyaml
lxml
